file changes and addition to camera and cube
