--- Assets/Scripts/NetworkCommandLine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d79817ce66a6c5e42930442993829ad4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
