+++ Assets/Scripts/NetworkTestPlayer.cs
@@ -1,0 +1,39 @@
+using Unity.Netcode;
+using UnityEngine;
+
+namespace NetworkTest
+{
+    public class NetworkTestPlayer : NetworkBehaviour
+    {
+        public NetworkVariable<Vector3> Position = new NetworkVariable<Vector3>();
+
+        public override void OnNetworkSpawn()
+        {
+            if (IsOwner)
+            {
+                Move();
+            }
+        }
+
+        public void Move()
+        {
+            SubmitPositionRequestServerRpc();
+        }
+
+        [Rpc(SendTo.Server)]
+        void SubmitPositionRequestServerRpc(RpcParams rpcParams = default)
+        {
+            transform.position = Position.Value = GetRandomPositionOnPlane();
+        }
+
+        static Vector3 GetRandomPositionOnPlane()
+        {
+            return new Vector3(Random.Range(-3f, 3f), 1f, Random.Range(-3f, 3f));
+        }
+
+        void Update()
+        {
+            transform.position = Position.Value;
+        }
+    }
+}
