+++ Assets/Scripts/NetworkTransformTest.cs
@@ -8,8 +8,8 @@ public class NetworkTransformTest : NetworkBehaviour
     {
         if (IsServer)
         {
-            float theta = Time.frameCount / 10.0f;
-            transform.position = new Vector3((float)Math.Cos(theta), 0.0f, (float)Math.Sin(theta));
+            // float theta = Time.frameCount / 10.0f;
+            // transform.position = new Vector3((float)Math.Cos(theta), 0.0f, (float)Math.Sin(theta));
         }
     }
 }
